+++ realm/realm-annotations-processor/src/test/resources/io/realm/NullTypesRealmProxy.java
@@ -1081,6 +1081,10 @@ public class NullTypesRealmProxy extends NullTypes
     }
 
     public static NullTypes copyOrUpdate(Realm realm, NullTypes object, boolean update, Map<RealmObject,RealmObjectProxy> cache) {
+        if (((RealmObject) object).realm != null && ((RealmObject) object).realm.threadId != realm.threadId) {
+            throw new IllegalArgumentException(
+                    "Objects belong to another Realm instance in other threads cannot be copied into this Realm instance.");
+        }
         if (((RealmObject) object).realm != null && ((RealmObject) object).realm.getPath().equals(realm.getPath())) {
             return object;
         }